Ignore build and test artifacts
This commit updates the .gitignore file to ignore the build artifacts of an Angular project. The directories dist/, tmp/, and out-tsc/ are now ignored. Additionally, the .angular/ directory is also added to the ignore list. The commit also updates the ignore list to include the bower_components/ directory and the coverage/ directory used for testing.